SALIENCY_DISAGREEMENT/SOURCE/requirements.txt
albumentations==0.5.2
captum==0.6.0
matplotlib==3.7.1
numpy==1.25.0
pandas==2.0.3
Pillow==9.5.0
scipy==1.11.1
seaborn==0.12.2
scikit-image==0.21.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
filelock==3.12.2
fonttools==4.40.0
idna==3.4
imageio==2.31.1
importlib-resources==5.12.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
lazy_loader==0.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
opencv-python-headless==4.7.0.72
packaging==23.1
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
requests==2.31.0
scikit-learn==1.2.2
six==1.16.0
sympy==1.12
threadpoolctl==3.1.0
tifffile==2023.4.12
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
zipp==3.15.0
